lib: Add initial library source code
Add the initial source code for the new ti-rpmsg-char library.
This library is a simple utility library designed to provide
ease of usage with the Linux kernel rpmsg-char driver to exchange
messages to a rpmsg device published by a remote processor.

The library provides the following 4 API and 2 headers for
applications:
API:
 1. rpmsg_char_init()
 2. rpmsg_char_exit()
 3. rpmsg_char_open()
 4. rpmsg_char_close()

Headers: ti_rpmsg_char.h and rproc_id.h

The rpmsg_char_init() and rpmsg_char_exit() functions are used
to initialize and finalize the library. The rpmsg_char_open()
is used to create and open a local rpmsg endpoint device. The
opened file handle can be used to read and write messages to
the corresponding rpmsg device. The rpmsg_char_close() is used
to close and destroy a previously created rpmsg endpoint device.

Signed-off-by: Suman Anna <s-anna@ti.com>