examples: Add rpmsg_char_simple example
Add a very basic example application demonstrating the usage
of the rpmsg_char API provided by the ti-rpmsg-char library.
The application is a single-threaded application, communicating
with a single rpmsg device on a remote processor using one local
endpoint device.

The example is added in a new 'examples' directory. An README file
is also added in the examples folder with the usage instructions.

The examples folder is not built automatically using top-level make,
and requires a specific command to be built. The top-level README
is also updated accordingly.

Signed-off-by: Suman Anna <s-anna@ti.com>