lib: Add initial signal handling code
The rpmsg char driver follows a peculiar design of using a control
device to create local endpoint devices, but using the endpoint device
itself to destroy the endpoint device. As such, any abnormal termination
of a process leaves stale endpoint devices (cleaned up only when rpmsg
char driver is uninstalled).

Add support to the library code to perform the needed cleanup of these
stale devices with unexpected terminations of the applications. Note
that there is no cleanup with SIGKILL (process terminated with kill -9)
as this signal cannot be captured by a running process.

NOTE:
This is not an ideal design, and actually needs to be ideally handled in
the applications themselves, as they might need to execute additional
cleanup code.

Signed-off-by: Suman Anna <s-anna@ti.com>