lib: Add support for J784S4 SoCs
J784S4 SoC has three dual core R5F sub systems in main domain and
four C71 DSPs in main compute cluster. The library is updated
accordingly to support J784S4 SoCs.

The library version is bumped to 0.5.0 to account for the addition of
the J784S4 SoCs to the library.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>