lib: Add support for AM62x Wkup R5F core
The TI AM62x SoCs have an R5F core in WKUP domain. This R5F is typically
used for running Device Manager software, but along with that, the R5F
core can also be used as remote processor for running custom application
(for e.g AUTOSAR based safety application) and communicate with the linux
host using virtio based rpmsg communication.

Add support for this R5F on AM62x SoCs in ti_rpmsg_char library. The
library version is bumped to 0.5.1 to account for these updates to
to the AM62x SoC support.

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>