lib: Add support for AM62Ax SoCs
AM62Ax SoCs have one R5F core in WKUP domain, one R5F core in MCU domain.
And in addition it has a C71 DSP in main domain. The library is updated
accordingly to support AM62Ax SoCs.

The library version is bumped to 0.5.1 to account for the addition of the
AM62Ax SoCs to the library.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Devarsh Thakkar <devarsh@ti.com>
Signed-off-by: Jai Luthra <j-luthra@ti.com>