lib: Add support for AM62Px SoCs
The TI AM62Px SoCs have an R5F core in WKUP domain. This R5F is typically
used for running Device Manager software, but along with that, the R5F
core can also be used as remote processor for running custom application
(for e.g AUTOSAR based safety application) and communicate with the linux
host using virtio based rpmsg communication. And in addition, the AM62Px
SoC have a R5F core in MCU power domain as well. This R5F core is used as
a remote processor and virtio based rpmsg communication is used for IPC.

Add support for these R5F on  AM62Px SoCs in ti_rpmsg_char library. The
library version is bumped to 0.6.3 to account for these updates to
support AM62Px SoCs.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>