Replace read bytes from 256 --> 496
The remote core can send up to 496 bytes per RPMsg. So there is no reason to
read back 256 bytes. The 2 valid approaches are:

1) read the max number of bytes that can be sent (496). This will work in all
usecases

2) If Linux knows the size that the message will be, then Linux can read back
only the expected number of bytes instead of the max number of bytes

We will use option 1 since this example could be modified for unknown usecases.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>