benchmark: update max runs = 1 trillion, max latency = 9sec
Realistically a customer will not run 450 trillion tests. cut the max
supported number of runs to 1 trillion (which would take months to run,
even if the round trip latency was negligible and the test was just limited
by the Linux interrupt response time, as per average cyclictest latencies).

This allows us to bucketize latencies in 1 usec increments of up to 9 seconds.

Also optimize the calculations and histogram data by using the worst-case
latency, instead of reading and writing all 9,000,000 potential latency values.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>